move lead on_trash to override
